Changes in compute_velocity function fro FreelySwimmingExperiment
The compute_velocity function is no longer cached, and its output is no
longer appended to the behavior_log. Instead, it returns a separate
pd.DataFrame. Made some small changes to fit that in the tests and the
documentation notebook, and added the latter to the docs/index.rst file.